Google Gemini projects/Gemini-Vision-App/requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
